dealt with issues spotted by Clang 3.3

+++ Source/GSGormLoading.m
@@ -370,7 +370,7 @@ static NSString *GSInternalNibItemAddedNotification = @"_GSInternalNibItemAddedN
   else
     {
       [NSException raise: NSInternalInconsistencyException
-		   format: @"Unable to write GSNibContainer version #%d.  GSNibContainer version for the installed gui lib is %d.", version, GNUSTEP_NIB_VERSION];
+		   format: @"Unable to write GSNibContainer version #%ld.  GSNibContainer version for the installed gui lib is %d.", (long)version, GNUSTEP_NIB_VERSION];
     }
 }
 
@@ -451,7 +451,7 @@ static NSString *GSInternalNibItemAddedNotification = @"_GSInternalNibItemAddedN
   else
     {
       [NSException raise: NSInternalInconsistencyException
-		   format: @"Unable to read GSNibContainer version #%d.  GSNibContainer version for the installed gui lib is %d.  Please upgrade to a more recent version of the gui library.", version, GNUSTEP_NIB_VERSION];
+		   format: @"Unable to read GSNibContainer version #%ld.  GSNibContainer version for the installed gui lib is %d.  Please upgrade to a more recent version of the gui library.", (long)version, GNUSTEP_NIB_VERSION];
     }
 
   return self;
